requirements.txt
Click==8.1.3
Flask==2.1.0
itsdangerous==2.0
Jinja2==3.1.2
MarkupSafe==2.0
Pillow==9.0.1
pytesseract==0.2.7
Werkzeug==2.0
